refactor data.go
* Created new direcory internal for internal functions
* created database package for interacting with data
* moved data.go and data_test.go to database directory
