+++ internal/database/data_test.go
@@ -1,4 +1,4 @@
-package main
+package database
 
 import "testing"
